+++ packages/freezed/test/integration/json.dart
@@ -1,5 +1,3 @@
-import 'package:meta/meta.dart';
-import 'package:json_annotation/json_annotation.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 part 'json.freezed.dart';