+++ packages/freezed/test/integration/nullable.dart
@@ -1,0 +1,35 @@
+import 'package:freezed_annotation/freezed_annotation.dart';
+
+part 'nullable.freezed.dart';
+
+@freezed
+abstract class RequiredPositional with _$RequiredPositional {
+  factory RequiredPositional(int a) = _RequiredPositional;
+}
+
+@freezed
+abstract class NullableRequiredPositional with _$NullableRequiredPositional {
+  factory NullableRequiredPositional(@nullable int a) =
+      _NullableRequiredPositional;
+}
+
+@freezed
+abstract class Positional with _$Positional {
+  factory Positional([int a]) = _Positional;
+}
+
+@freezed
+abstract class Named with _$Named {
+  factory Named({int a}) = _Named;
+}
+
+@freezed
+abstract class RequiredNamed with _$RequiredNamed {
+  factory RequiredNamed({@required int a}) = _RequiredNamed;
+}
+
+@freezed
+abstract class NullableRequiredNamed with _$NullableRequiredNamed {
+  factory NullableRequiredNamed({@nullable @required int a}) =
+      _NullableRequiredNamed;
+}