+++ packages/freezed/test/integration/single_class_constructor.dart
@@ -1,5 +1,5 @@
-import 'package:meta/meta.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:meta/meta.dart';
 
 part 'single_class_constructor.freezed.dart';